requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.4.0
